[#110] refactor : validateLetter 유틸 파일로 분리 및 재사용
