+++ src/pages/friend/FriendDraftPage.tsx
@@ -12,11 +12,7 @@ import { useDailyQuestion } from '@/hooks/letters/useDailyQuestion';
 import { useGlobalToast } from '@/components/toast/ToastProvider';
 import LoadingPage from '../system/LoadingPage';
 import { Button } from '@/components/common/Button';
-
-const LIMIT = {
-  TITLE: { MIN: 3, MAX: 20 },
-  CONTENT: { MIN: 1, MAX: 500 },
-} as const;
+import { validateLetter } from '@/utils/validateLetter';
 
 export default function FriendDraftPage() {
   const { data, isLoading, isError, refetch } = useDailyQuestion();
@@ -36,17 +32,6 @@ export default function FriendDraftPage() {
 
   const location = useLocation();
   const friendName = (location.state as { friendName?: string })?.friendName;
-
-  const validate = (title: string, content: string) => {
-    if (title.length < LIMIT.TITLE.MIN) return `제목을 ${LIMIT.TITLE.MIN}자 이상 입력해주세요.`;
-    if (title.length > LIMIT.TITLE.MAX)
-      return `제목은 최대 ${LIMIT.TITLE.MAX}자까지 입력할 수 있어요.`;
-    if (content.length < LIMIT.CONTENT.MIN) return '내용을 작성해 주세요!';
-    if (content.length > LIMIT.CONTENT.MAX)
-      return `내용은 최대 ${LIMIT.CONTENT.MAX}자까지 입력할 수 있어요.`;
-
-    return null;
-  };
 
   // questionId 저장
   useEffect(() => {
@@ -72,7 +57,7 @@ export default function FriendDraftPage() {
 
     const title = draft.title.trim();
     const content = draft.content.trim();
-    const errorMsg = validate(title, content);
+    const errorMsg = validateLetter(title, content);
 
     if (errorMsg) {
       showToast(errorMsg, 'error');
