+++ src/pages/letter/OtherDraftPage.tsx
@@ -13,11 +13,7 @@ import { useEffect } from 'react';
 import LoadingPage from '../system/LoadingPage';
 import { Button } from '@/components/common/Button';
 import { useThreadFlowStore } from '@/stores/letterContext';
-
-const LIMIT = {
-  TITLE: { MIN: 3, MAX: 20 },
-  CONTENT: { MIN: 1, MAX: 500 },
-} as const;
+import { validateLetter } from '@/utils/validateLetter';
 
 const OtherDraftPage = () => {
   const { data, isLoading, isError, refetch } = useDailyQuestion();
@@ -39,17 +35,6 @@ const OtherDraftPage = () => {
 
   // TODO : 남은 편지 횟수 처리 필요
   const letterLeft = 4;
-
-  const validate = (title: string, content: string) => {
-    if (title.length < LIMIT.TITLE.MIN) return `제목을 ${LIMIT.TITLE.MIN}자 이상 입력해주세요.`;
-    if (title.length > LIMIT.TITLE.MAX)
-      return `제목은 최대 ${LIMIT.TITLE.MAX}자까지 입력할 수 있어요.`;
-    if (content.length < LIMIT.CONTENT.MIN) return '내용을 작성해 주세요!';
-    if (content.length > LIMIT.CONTENT.MAX)
-      return `내용은 최대 ${LIMIT.CONTENT.MAX}자까지 입력할 수 있어요.`;
-
-    return null;
-  };
 
   // questionId 저장
   useEffect(() => {
@@ -75,7 +60,7 @@ const OtherDraftPage = () => {
 
     const title = draft.title.trim();
     const content = draft.content.trim();
-    const errorMsg = validate(title, content);
+    const errorMsg = validateLetter(title, content);
 
     if (errorMsg) {
       showToast(errorMsg, 'error');
